+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-03-26
+
 ### Added
 
 - VCS checkin handler to check data files before commit
@@ -13,8 +15,12 @@
 - Used virtual file API to read and write files
 - Persisted settings now use a `SerializablePersistentStateComponent`
 
-## 0.1.0 Initial Release
+## [0.1.0] Initial Release
 
 ### Added
 
 - Property file sorting and formatting
+
+[Unreleased]: https://github.com/Kaemmelot/intellij-data-files-sorter/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/Kaemmelot/intellij-data-files-sorter/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/Kaemmelot/intellij-data-files-sorter/commits/v0.1.0